% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{builtin_dist_funs}
\alias{builtin_dist_funs}
\title{All built-in feature dissimilarity functions}
\format{
An object of class \code{list} of length 5.
}
\usage{
builtin_dist_funs
}
\description{
To be used in \code{\link{run_rafs}}.
}
\details{
See also \code{\link{default_dist_funs}}.
}
\keyword{datasets}
