% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{cor_dist}
\alias{cor_dist}
\title{Feature dissimilarity based on Pearson's Correlation (cor)}
\usage{
cor_dist(relevant_train_data, train_decision = NULL, seed = NULL)
}
\arguments{
\item{relevant_train_data}{input data where columns are variables and rows are observations (all numeric); assumed to contain only relevant data}

\item{train_decision}{decision variable as a binary sequence of length equal to number of observations}

\item{seed}{a numerical seed}
}
\value{
A matrix of distances (dissimilarities).
}
\description{
To be used as one of the \code{dist_funs} in \code{\link{run_rafs}}.
}
