% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_seeded_folds}
\alias{create_seeded_folds}
\title{Create seeded folds}
\usage{
create_seeded_folds(decision, k, seed)
}
\arguments{
\item{decision}{decision variable as a binary sequence of length equal to number of observations}

\item{k}{number of folds for cross validation}

\item{seed}{a numerical seed}
}
\value{
A vector of folds. Each fold being a vector of selected indices.
}
\description{
A utility function used in RAFS but useful also for external cross-validation.
}
