% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{default_dist_funs}
\alias{default_dist_funs}
\title{Default feature dissimilarity functions}
\format{
An object of class \code{list} of length 3.
}
\usage{
default_dist_funs
}
\description{
As used in \code{\link{run_rafs}}.
}
\details{
The default functions compute:
Pearson's correlation (\code{cor}: \code{\link{cor_dist}}),
Variation of Information (\code{vi}: \code{\link{vi_dist}}) and
Symmetric Target Information Gain (\code{stig}: \code{\link{stig_dist}}).

These functions follow a similar protocol to \code{\link{default_fs_fun}}.
They expect the same input except for the assumption that the data passed in is relevant.
Each of them outputs a matrix of distances (dissimilarities) between features.

See also \code{\link{builtin_dist_funs}}.
}
\keyword{datasets}
