% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{default_hclust_methods}
\alias{default_hclust_methods}
\title{Default hclust methods}
\format{
An object of class \code{character} of length 4.
}
\usage{
default_hclust_methods
}
\description{
As used in \code{\link{run_rafs}} to call \code{\link{hclust}}.
}
\keyword{datasets}
