% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_run_id}
\alias{get_run_id}
\title{Generate CV run identifiers}
\usage{
get_run_id(seed, k, i)
}
\arguments{
\item{seed}{a numerical seed}

\item{k}{number of folds for cross validation}

\item{i}{current fold number (1 to \code{k})}
}
\value{
A string with the run identifier.
}
\description{
A utility function used in RAFS to generate cross validation run identifiers, thus useful also for external cross-validation.
}
