% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{stig_dist}
\alias{stig_dist}
\title{Symmetric Target Information Gain (STIG) computed directly}
\usage{
stig_dist(relevant_train_data, train_decision, seed)
}
\arguments{
\item{relevant_train_data}{input data where columns are variables and rows are observations (all numeric); assumed to contain only relevant data}

\item{train_decision}{decision variable as a binary sequence of length equal to number of observations}

\item{seed}{a numerical seed}
}
\value{
A matrix of distances (dissimilarities).
}
\description{
To be used as one of the \code{dist_funs} in \code{\link{run_rafs}}.
}
\details{
This function computes the STIG metric directly from the data, maximising it over 30 discretisations.
}
